sync-service/requirements.txt
# Government Portal Sync Service Dependencies
# Python 3.10+

# Google Drive API
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Database
psycopg2-binary==2.9.9

# File parsing
markdown==3.5.1
pandas==2.1.4
openpyxl==3.1.2

# Utilities
python-dotenv==1.0.0
